test(research): split PDF equation ownership coverage
